refactor(markdown): Change CodeBlock to 'markdown-blocks' and move CodeBlock component to its own file
